lib: support big endian devices
With big endian devices, the symbols are stored in big endian order. This is
fine if the header is getting loaded onto the target, however the header is
only for use with the host system, which is almost alays little endian.

Ensure each field of the header is correctly swapped by adding `.to_le()` to
each field. Additionally, ensure all fields are primitive integer types, as
byte swapping is undefined for Rust types.

Signed-off-by: Sean Cross <sean@xobs.io>